controller: volume: Don't update state to attaching if it's only healing
When the engine is running, ReconcileVolumeState will still need to make sure
everything is in the correct state so it will nonetheless make sure every good
replicas are running and engine is running.

If we're during a rebuild process, we will need to start new replicas in this
process. But it doesn't mean the volume is `attaching`. It's already attached.