controller: engine: Remove the replica if rebuilding failed
If the removal succeed, the manager will try to rebuild with the same replica
again.

But if the removal failed, we will notify the user and let user decided what to
do. We don't want to automatically mark the replica as failed then we may cause
a chain reaction of starting numerous new replicas. User can delete the replica
failed rebuild manually.