controller: engine: Fix monitor using the old engine client after upgrade
The bug is triggered by:
1. The engine has been upgraded
2. No more reference to the old image

The old image will be cleaned up in 1 hour. But the monitor of engine controller
still using it instead of upgrade to the new image.

Use the current image version matched client everytime for the engine
controller.

Test instruction:
1. Upgrade to a new image.
2. Make sure there is no other reference for the old image
3. Wait more than a hour (or change the parameter by compiling a new binary).
Confirm the old image was cleaned up.