controller: instance: Fix sometime status.Started and CurrentImage wasn't set
Previous Started and CurrentImage was only set after we successfully create the
pod. But the problem with that is if the update of instance object (engine or
replica) failed due to conflict, Started and CurrentImage will never be set
again, cause major malfunction later.

Now Started will be set whenever the instance is suppose to be running and pod
exists. CurrentImage will be set along with status.IP, and only set once when
it's empty.