Fix race condition in replica scheduler
ReplicaScheduler currently depends on using the volume replica list for
scheduling. But the creation operation in data store wasn't a barrier
operation, so it won't guarantees the result of creation will show
immediately in the following listing operation. As the result, sometimes
the replica just had been created won't be taken into scheduling
consideration, result in multiple replicas could be scheduled to the
same node even there are other nodes available.