controller: volume: Update scheduled condition in the ReconcileVolumeState
Because create replica has a side effect, and if the replicas are created but
the update to the condition failed, the `scheduled` condition will never be
set again.

It's fine with `scheduled = false` though since it will go through the same
code path to replenish the replicas anyway next time.