controller: node: only update node once
The controller doesn't hold the reference to the updated node object, and there
are 3 updates may happen in a loop. So if any one of them success, the revision
of the object will change, result in all the following update fails.

Now the node object should be updated only in one place, and the reference of
the object will be used through the reconciliation loop.

https://github.com/rancher/longhorn/issues/162