controller: volume: mark the replicas on down nodes as failed when attaching
Otherwise we will wait for it forever.

And we will try to delay it as much as possible. So we won't declare it as
failed until the volume attached again.

Also, the logic to determine if a node is down has been changed a bit. Now we
don't depend on the exact reason on why node is not ready.