util: Don't combine stderr to output
The failed test exposed one bug:

> Server Error : error listing backups for volume
> 'longhorn-testvol-qni5x6': error parsing BackupTarget:
> \ntime="2019-03-21T00:57:47Z" level=error msg="{\n AcceptRanges:
> \"bytes\",\n Metadata: {\n\n }\n}" pkg=s3
> \n{\n\t"longhorn-testvol-qni5x6": {\n\t\t"Name":
> "longhorn-testvol-qni5x6",\n\t\t"Size": "0",\n\t\t"Created":
> "",\n\t\t"LastBackupName": "",\n\t\t"SpaceUsage":
> "0",\n\t\t"Messages": {\n\t\t\t"error": "cannot find
> backupstore/volumes/02/1b/longhorn-testvol-qni5x6/volume.cfg in
> backupstore"\n\t\t}\n\t}\n}\n: invalid character 'i' in literal true
> (expecting 'r')\n{'status': 500, 'code': u'Server Error',

What's shown here is there is an additional log line somehow triggered
before the output. Though the log line is written into stderr, it was
caught by the CombinedOutput() and treated as json in the end, which
caused the error.

The reason of why the log line shows up now hasn't been determined, but
the assumption that the command's output (including stderr) should be
treated as the return result is flawed. We should only took stdout
instead.

https://github.com/rancher/longhorn/issues/437