api: Add LastBackup and LastBackupAt field in the volume object
Those two object would indicate what's the last backup and what's the
time the backup was made.

Currently the fields would be set when a new backup has been made, from
UI or using recurring backup feature.

The field would be refreshed when a backup get removed. Current it would
be unset when the last backup get deleted.

https://github.com/rancher/longhorn/issues/459