Support Bundle Enhancement
Earlier the support bundle download was a single GET request and it was a blocking call where the request was forwarded to the manager and when the download completes it returns the compressed zip file back to the caller. This was causing timeout if the content is huge.
In order to fix this issue, the request is split into 3 requests:
1. POST request to initiate a support bundle download and the backend returns the nodeid of the manager that's performing this operation
2. GET request to check the status of download. This returns true or false based on the progress of the download.
3. POST request to download the compressed file.