Update EngineSpec OwnerID when the existing OwnerID goes down
The engine spec owner ID is not updated when the node that is referred by the current ownerID goes down. Controllers running in other nodes wouldn't pick up as the ownerID value is not empty.
Hence the following code changes explicitly check if the node is down and then set owner ID to the current controller ID.