Controller changes for Restore
Earlier, the replica controller was doing restore when bringing up the replicas
but now it's executed by volume controller after all the replicas have come up.
Specifically, as soon as a restoration volume is created, it's automatically
attached and detached and in this window, the backup restore engine command is
executed to restore all the replicas.