controller: Set InstanceManagerName when instance is starting/running
For the stage that createInstance() is called but the instance state
of the InstanceManager hasn't been updated, we prefer not to keep
InstanceManagerName since the code is complicated. Then we need to
make sure InstanceManagerName won't be changed in this stage.

Now we are using node ID and engine image to fetch InstanceManager.
Since the node ID shouldn't be modified, we just need to make sure
engine image won't be modified in that stage. The way to modify
engine image is engine upgrade. But in that stage, volume is
attaching then both live upgrade and offline upgrade are not
available. Hence engine.Spec.EngineImage won't be modified and the
InstanceManager shouldn't be changed.

Signed-off-by: Shuo Wu <shuo@rancher.com>