Remove the field `Readiness` for Node status
This field is introduced by the PR:
https://github.com/longhorn/longhorn-manager/pull/370

Since there can be multiple engine images in the Longhorn system,
we cannot mark the node as state `deploying` and disable all volume
creation if one of the engine images is not ready.

We plan to use engine image state later to indicate if the relatad
instance managers are `running`.

Signed-off-by: Shuo Wu <shuo@rancher.com>