controller: Move instance manager creation and deletion to engine image controller
Deleting engine image will lead to implicit deletion of instance
manager. But instance manager is always created by node controller.
This means that there are two controllers managing the life cycle
of instance manager, which is wrong.

The corner case:
1. The engine image is being deleted.
2. Node controller and engine image controller are triggered at
   the same time.
3. Engine image controller deletes engine image and related instance
   managers.
4. Then node controller creates new instance managers based on the
   deleted engine image.
5. Since the engine image is gone, there is no way to clean up those
   wrongly created instance managers.

Now we should have the same controller in charge of creation and
deletion of the instance manager.

Signed-off-by: Shuo Wu <shuo@rancher.com>