controller: engine: Cover instance manager running with the correct owner
It doesn't make sense to limit the instance manager's update to the
owner of instance manager controller:
1. During node down, the instance manager controller's owner will change
to the manager on another node, but nodeID remains pointed to the
original node since it's still managing the instances on the original
node.
2. The engines should be updated by the engine owner. Any engine
controller may temporarily own the engine on another node if that node
is down. So there is no way to know if the engine controller own's the
engine of this instance manager or not.

Signed-off-by: Sheng Yang <sheng.yang@rancher.com>