controller: instance manager: Fix high cpu utilization caused by health probe
The wrong configuration results in skyrocketed CPU usage on instance
manager pods. The liveness and readiness probe intervals are set to one
second, which is pushing the system to check health way too frequently.
Being an `exec` type of probe doesn't help too.

This patch set liveness probe interval to 5 seconds, with the initial
delay of 3 seconds.

This patch also removes readiness probe since instance mangers are not behind a
Kubernetes Service.

Longhorn #767

Signed-off-by: Sheng Yang <sheng.yang@rancher.com>