controller: Unset the instance manager name if we cannot find the related instance or instance manager
If the instance process somehow gets crashed and the instance
handler directly unset the field `instanceManagerName`, then
longhorn manager won't send request to the related instance manager
to  clean up the process since it doesn't know which the instance
manager should be used.

Later if we detached and re-attach the volume, the instance manager
will find the related process already exists with `error` state
hence it will do nothing for the re-attach. This volume cannot become
attached anymore.

Longhorn #789

Signed-off-by: Shuo Wu <shuo@rancher.com>