controller: Check ownerID before updating volume's kubernetes status for PV deletion in kubernetes controller
cleanupForPVDeletion() will clean up the volume's kubernetes status
then update the volume if the volume's PV is deleted. In order to
avoid conflict of volume update, we need to check ownerID for the
volume.
Since this function will check if the PV exists and decide if the
main reconciliation needs to be executed, it is triggered before the
main reconciliation. Then we need to add this separate ownerID check
for it.

Signed-off-by: Shuo Wu <shuo@rancher.com>