controller: Skip updating instance manager with deletion timestamp during cleanup
The update function will automatically add finalizer for the objects,
which is not allowed for deleting objects.

For the instances in the deleting instance manager, we don't need to
worry about them. They will become state error once the instance
manager is gone. Hence it's safe to skip the update during instance
manager cleanup.

Signed-off-by: Shuo Wu <shuo@rancher.com>