controller: Set volume fault state according to the replica status
Rather than set it manually in manager.

Also move the detach code in the `salvage` call to earlier since
updating replicas will result in volume object changes (though it's the
status not the spec), then the write to `nodeID` to detach the volume
will result in a conflict (since status changed) and a failed API call.

Signed-off-by: Sheng Yang <sheng.yang@rancher.com>