Update Longhorn related PV/PVC annnotations for new CSI provisioner
1. The new external CSI provisioner will use driver name as its
provisioner name. Previously, the provisioner name is different
from driver name, which is `rancher.io/longhorn` and does not
follow the driver naming schema. Hence we cannot deploy a
compatible provisioner and a related CSI driver with old provisioner
name to handle the existing volumes.

2. The external CSI provisioner uses the annotations of PVC/PV to
find the matching CSI driver. Hence we can updating the annotations
for Longhorn related PVC/PV so that the new provisioner is able to
verify and handle the existing volumes.

Longhorn #347

Signed-off-by: Shuo Wu <shuo@rancher.com>