controller: Remove the logic to creating the replica directory when adding disks
It's done in the controller as a part of generateDiskConfig now.

We kept the creation call in the create default disk using
`default-data-path` since we might use the root disk for the default
data path.

Also, we're not going to automatically recreate replica directory, so
the test case in  #1147 will be removed.

Signed-off-by: Sheng Yang <sheng.yang@rancher.com>