manager: Remove unecessary wait during the PV or PVC creation
Longhorn needs to wait for PV available only before calling
the PVC creation API. And waiting for the status update after
calling the Kubernetes API is unnecessary.

Notice that the PV wait may trigger the similar issue later. But
we don't want to require users to input PV name for this API since
it's not intelligent enough. This PV wait is actually a compromise.

Longhorn #1092

Signed-off-by: Shuo Wu <shuo@rancher.com>