Change csi driver to let volume api make attach/detachment decisions
Previously the csi driver decided whether attachment/detachment was
necessary, but it didn't actually have all required information to make
this choice. This could then lead to a situation in which the csi driver
told kubernetes that attachment was successful but in reality longhorn
never attached the volume. Once that case happens the Node.PublishVolume
will continuously fail, but because kubernetes thinks the volume is
already attached it will not try to call Controller.PublishVolume again.

After this change we will always let the longhorn manager make the
decision whether attach/detachment is necessary so that the csi driver
(kubernetes) and longhorn cannot end up in a situation where they are
out of sync in regard to a volume attachment.

The longhorn manager will now also return successfully when a volume
attach/detachment is requested for a volume that is already in the
correct requested state.

Longhorn #820
Longhorn #159

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>