Add additional csi driver error conditions for ControllerPublishVolume
We now return the correct error code `FailedPrecondition` for a volume
that is already attached to another node. We now also return `NotFound`
for the case where a node is `NotReady` which is the case if longhorn
detects a NodeFailure. Both of these are the expected error codes and
will lower the time it takes Kubernetes to try to attach the volume to
another another node.

Longhorn #820

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>