Fix csi driver call detach volume only if attached to the requested node
Since the manager api does not receive a nodeId as part of the detach
request a stray detach request from the csi driver would detach the
volume from a node that it's currently correctly working on.
This stray detach call could happen when a downed node comes back to
life since kubernetes can now cleanup the stuck pods which would lead to
the deletion of existing volume attachments, if they have not been
deleted already.

Longhorn #820

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>