controller: Remove redundancy for engine/replica deletion
During the deletion test, I found the implementation of the
replica controller was different from that of the engine controller.
And there were some redundancy for the deletion implementation.

This commit will remove the redundancy, as well as make the replica
deleteion implementation become more straightforward and be
consistent with the engine controller.

Signed-off-by: Shuo Wu <shuo@rancher.com>