controller: Fix the deadlock during the engine monitor stopping
1. The deadlock happenes in the engine monitor sync loop: the
signaling operation in the sync loop may be blocked by the existing
singal in the channel, then the following singal receiving in
function wait.Until() is blocked, too. That's the deadlock.
2. Actually the implementation is weird. The callee (engine monitor
sync loop) here is using channel to talk with the caller (wait.Until).
Instead, we can remove the function wait.Until (which relies on the
stop channel to stop the loop) and directly return a boolean after
each sync loop. As a result, the sync function itself has no blocking
operation and the deadlock can be solved.
3. Since there is only one goroutine sending the signal to the channel
now, we can directly close the channel and remove the related entry
after the signal sending.
4. Add conflict retry count for the engine monitor sync loop, which
guarantees that the monitor stop process won't be blocked by the
possible endless conflicts.

Longhorn #1545

Signed-off-by: Shuo Wu <shuo@rancher.com>