controller: Clean up the instance manager when the node is disconnected
When the node is disconnected and the related IM pods is still
running, the engine processes on the nodes will wrongly mark all
replicas as `ERR` due to the ping timeout. Then the node and the
pods are back, the longhorn managers will update all replicas
based on the wrong engine process status. The volume will become
`Faulted`.

The solution here is that:
Directly do cleanup and return from the IM comtroller sync loop
for Unknown IM objects. And there shouldn't be Unknown IM objects
in following IM pod sync logic.

With this fix, all engines on the nodes will be considered as
crashed unexpectedly once the IM pods are back.

Longhorn #1545

Signed-off-by: Shuo Wu <shuo@rancher.com>