Retain compatible code for old engines restore, and update the API version
1. A new constant `MinCLIVersion` is introduced since the manager
now will apply different invocations based on the engine process
API version. Then the way to identify incompatible engine image
should be updated.
2. Considering the fact that the restore status in the server will
be lost after the live upgrade(sync agent server restarts) and the
incremental restore relies on the the restore status in the new
workflow, the 1st restore after the live upgrade is a full restoration.
Hence it's better to support incremental restore for the old DR
volumes. And we may need to add a warning for the users who decides
to do live upgrade for the DR volumes.

Longhorn #1279

Signed-off-by: Shuo Wu <shuo@rancher.com>