controller: Support rebuilding for restore/DR volumes
1. For new restore/DR volumes, the auto detachment won't be applied
if the rebuilding is not down.
2. For old restore/DR volumes, we need to make sure that the volume
can be automatically detached even if the volume is degraded.
3. Since the rebuilding replicas won't sync files from running
replicas, the engine controller can rebuild all replicas for restore/
DR volumes simultaneously.
4. The rebuilding replicas will lead to the last restored backup
being inconsistent, we need to unset engine.Status.LastRestoredBackup
and update the restore condition when the rebuilding starts.
5. Make sure the frontend is always disabled when restore/DR volume
haven't been detached automatically.
6. Update the unit tests

Longhorn #1279

Signed-off-by: Shuo Wu <shuo@rancher.com>