vendor: fix nil pointer for Leader Election Record
This merges in a fix for the LeaseSpecToLeaderElectionRecord function
which isn't available until Kubernetes v1.17 which helps prevent a nil
pointer reference when LeaseDurationSeconds is set. This fix is based on
the changes in this commit:

kubernetes/kubernetes@7907b29551c7ef87bbe398ac02836b4c87246d3d

This patch can be dropped once Longhorn Manager switches to using the
Kubernetes v1.17 API, which contains this fix.

Signed-off-by: Michael William Le Nguyen <michael@mail.ttp.codes>
Signed-off-by: Phan Le <phan.le@rancher.com>