Make sure the rebuild is disabled during the expansion
Need to consider the following corner case:
1. The DR volume need to expand the size before doing the next inc
restore.
2. There is a rebuilding replicas just launched with mode WO.
3. The rebuilding replica will get stuck in mode WO without
finishing the restore. But the whole volume need to do expansion
before the restore. And the expansion cannot be started if there
are rebuilding replicas. Deadlock!

Hence we need to mark all rebuilding replicas as ERROR and disable
the replica replenishment before the expansion complete.

Longhorn #1279

Signed-off-by: Shuo Wu <shuo@rancher.com>