Refactor the auto attach/detach and restore and expansion in volume controller
1. Stop checking the restore condition before updating any fields.
The restore condition is used for presenting the current restore
status only.
2. Update the restore condition based on the final attachment state.
3. Use separate function to handle auto attachment/detachment logic.
4. Rename `volume.Status.InitialRestorationRequired` to
`volume.Status.RestoreRequired`. It will indicate if the
restore/DR volume finishes the restore. It's unset only when the
restore/DR volume is available/has been activated.
5. Avoid using `e.Spec.RequestedBackupRestore` to verify if a volume
is restore/DR volume. Use `volume.Status.RestoreRequired`
instead.
6. Blindly set `e.Spec.RequestedBackupRestore` when the restore is not
done. And blindly unset it when the volume is not attached.
7. Simplify the conditions for auto attachment/detachment and fix some
issues in the implementation.

Longhorn #1279

Signed-off-by: Shuo Wu <shuo@rancher.com>