controller, deploy, scheduler, types, api, manager: support Data Locality option to always keep a replica local to the engine
Give user option to keep a local replica to be on the same node
as volume engine. If a volume is attached to a node that has no
replica, Volume controller will start rebuilding the replica
on the node after the volume is attached. Once the rebuilding
process is done, it will remove one of the other replicas to
keep the replica count as specified.

If Longhorn cannot schedule a local replica on to the same node as
engine, it will not stop the volume. In other words, keeping
a local replica is best effort only, not a hard requirement.

Longhorn#1045

Signed-off-by: Phan Le <phan.le@rancher.com>