feature: Add AllowVolumeCreationWithDegradedAvailability
The setting allows user to create a volume with less than specified
number of replicas, so user can use the default Longhorn configuration
without issues in one or two nodes cluster.

It is recommended to disable the feature when using Longhorn in
production environment.

Implementation details:
1. It's the `scheduled condition` who blocks the volume being attached
if there is less than specified number of replicas.
2. This commit will check for the setting and set the condition to
`true` to allow the volume to be attached if there is at least one
replica has been scheduled successfully.

Note:
1. At creation, even if the `scheduled condition` is reset to true, the
additional `failed to schedule` replicas won't be cleaned up
automatically. The reason is we might still able to find another
suitable node before the first attaching. After the first attachment,
the volume will contain data, then there will be no way to add new
replicas when the volume is detached.

Longhorn #1701

Signed-off-by: Sheng Yang <sheng.yang@rancher.com>