Fix: volume remains attaching to the wrong node when it is detached unexpectedly and auto-reattached during a recurring backup
Change the logic of recurring backup job so that when the job
automatically attaches the volume to a node, it also sets volume's
LastAttachedBy field to be the job's name.

At the end of the recurring backup job, it will check to see if
the volume was attached by the job's name. If so, the job tries
to detach the volume.

Longhorn #1922

Signed-off-by: Phan Le <phan.le@rancher.com>