Increase default csi timeouts to 2 minutes 5s
Since we are polling for the longhorn volume state inside of the csi
plugin for a maximum of 2 minutes, it doesn't make much sense having
a default csi grpc timeout of 15 seconds, this will just lead to many
unnecessary csi call failures which keeps increasing the backoff time.

Having a higher timeout shouldn't be a problem since the operation should
complete before the timeout, but if it doesn't then the next csi call will
actually have an effect and retry the timed out operation.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>