Transfer share manager ownership when a node is being drained
We only transfer ownership if the node is unschedulable and the share
manageris in error state, since one shouldn't create new pods on an
unschedulable node. The reason we don't transfer ownership while the pod
is running, is that unschedulable only means that no new pods should be
created on this node.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>