Add volume data migration example job
This can be used to migrate data from other storage systems or the old nfs
provisioner. The job will restart up to 3 times in the case of failure of
the copy operation.

After successful completion the list of files, can be viewed in the logs
of the job pod. The job can be deleted once the operation succeeded.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>