Replace old rwx examples with an rwx nginx deployment
This uses the new native longhorn rwx support. This example comes with a
service and a persistent rwx volume claim for the nginx rwx deployment.
The deployment has 2 containers One for running nginx and another one for
continuously appending the current date and time to the index.html file.

The reclaim policy of the rwx pvc is based on the reclaim policy of the
longhorn storage class.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>