Use Volume attach/detach to determine migration intention
Instead of having the client call a migration start/finish/rollback api.
We can use the attach node and detach nodes to determine the current
migration intention, we only support migration for volumes that are set
to be migrateable and the access mode is rwx.

A volume can either be shareable (spec.Share==True) or migrateable
(Spec.Migrateable==True)

Longhorn #2111

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>

fixup! Use Volume attach/detach to determine migration intention

Had to revert the requested change, to error on non RWX access mode. Since
the kubevirt importer will use the same storage class in RWO mode. It makes
more sense to just unset the migratable flag and have that reflected in the
volume_context that is returned from the Volume Creation call.

This way the migratable storage class can be used to provision regular RWO
volumes as well. The alternative is to add a share=true param to the storage
class since that will force RWX access mode for longhorn, plus
migratable=true will lead to longhorn creating a migratable rwx volume.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>