Use Volume attach/detach to determine migration intention
Instead of having the client call a migration start/finish/rollback api.
We can use the attach node and detach nodes to determine the current
migration intention, we only support migration for volumes that are set
to be migrateable and the access mode is rwx.

A volume can either be shareable (spec.Share==True) or migrateable
(Spec.Migrateable==True)

Longhorn #2111

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>