Always check for volume ownership
We want to always check whether we should be the owner of a volume otherwise
we might end up doing an iteration while we are in the process of an ownership
transfer (v.Spec.NodeID != v.Status.CurrentOwnerID) this would lead to 2
controllers fighting about the same object and a couple of conflicts on
the volume updates. The real problem could be the updates that are run
on the dependent components (engine, replicas, etc) by the now no longer
valid owner of this volume.

Longhorn #2111

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>