Call attach/detach from the csi driver for all volume types
This allows the manager to reset the disableFrontend flag, for recently
expanded volumes. This also gives us the ability to track share usage for
different nodes, ie if we want to later implement nfs export restrictions.

Besides there is also a refactor for the csi driver included that simplifies
the ControllerPublishVolume and ControllerUnpublishVolume code.

Longhorn #2181

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>