Remove volume migrationNodeID from the api response
Instead of using this specific id the api consumer can use the controllers
list, same as when we want to see the case for a single attachment. Since
we don't expose the regular nodeID, this makes more sense not to expose
the migrationNodeID since that is internal implementation detail and might
change down the line.

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>