Fix volume expansion idempotency
Kubernetes doesn't support volume shrinking and the csi spec specifies
to return true in the case where the current capacity is bigger or equal
to the requested capacity that's why we return the volumeSize instead of
the requested capacity.

We also now poll as part of the ControllerExpandVolume to wait for the
actual volume expansion to be completed before returning success.

Longhorn #2181

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>