Fix backup monitor polling from each node at the same time
Currently the backupstore monitor poll is processed by every node in the
cluster, this is undesired since it leads to excessive S3 calls and cost.

This also has the side effect that each nodes backupstore monitor will
race against each other to try to update each volumes lastBackup field.
This leads to unnecessary conflicts and processing of the volumes.

Since the backup monitor is run on each node, but we only need a single
update we pick a consistent random ready node for each poll run. This is
accomplished by the following steps:
- sort the candidate list, this normalizes the list across nodes.
- use a time index to derive an index into the candidate list, this
  normalizes time differences across nodes.
- arbitrarily choose the pollInterval as your time normalization factor,
  since this also has the benefit of doing round robin across the at the
  time available candidate nodes.

Longhorn #1547

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>