Fixed engine image is never become deployed after upgrading
Because we recently change the labels for engine image CR, we need to
use the correct label selector to get the engine image daemonset pod.
The engine image daemonset pod shouldn't containts the label
`longhorn.io/managed-by: longhorn-manager` because it is not managed by
Longhorn manager. It is managed by the daemonset controller.

Longhorn #2399

Signed-off-by: Phan Le <phan.le@rancher.com>