Ownership transferring refactoring
1. When the manager pod on the owner node of the CR is missing, we switch
the ownerID of the CR to a different node
2. Don't cleaning up CRs base on the current ownerID. The convention is
that ownerID only indicates which node should watch the CR. We shouldn't
make any other logic depend on the ownerID. This way we can freely transfer
ownership
3. Make sure that when the engine/instance manager CR is transfered to
new ownerID, the old owner must stop monitoring them. Also, remove
engine/instance manager monitoring from the monitoring map of engine/instance
manager controller when stop monitoring the engine. This allow the old
owner to restart the monitoring when the CRs' owner is transfered back to it
4. Don't clear the engine status when switching ownership. The new owner
will take care of continuing to update the engine. This create a smooth
owner transferring.
5. Modify the isResponsibleFor() in engine, volume, engine image controller
so that the current owner can continue to be the owner if the preferred
owner doesn't have engine image deployed.
6. Make sure the replica/engine instance is able to stop when the IM is
in error state
7. The controller should at least look at each resource to determine
whether it's responsible for it or not, it's no the enqueues responsibility
to make that determination

Longhorn #2377
Longhorn #2329

Signed-off-by: Phan Le <phan.le@rancher.com>