Remove the schedule condition check from the Volume attach call
This condition is already covered by the readiness check in the detached
state. I identified some issues with the way the scheduling conditions are
handled in the volume controller, but they are outside of the scope for
v1.1.1 This is a surgical fix, to allow for rwx volumes to be used even if
one replica cannot be scheduled.

The share manager currently does not check for this condition, and will
always allow for volume usage even if not all replicas could be scheduled.
After creation, this is part of the problem, since the intent of the
scheduling failure is to keep the volume in the creation state, right now
the volume controller does not do this instead, it has some hacks for the
detached state.

I plan on looking into this, when working on the volume controller refactor.

Longhorn #2463

Signed-off-by: Joshua Moody <joshua.moody@rancher.com>