app: Increase the timeout for the http client
The default request timeout for recurring jobs is 10s, which is
definitely not enough. If multiple recurring jobs for a volume
start simultaneously (due to the mutex lock), or some steps take
several seconds, the timeout will fail the recurring jobs.

Increasing the timeout is not able to completely avoid duplicate
snapshots/backups if there are intensive snapshot-related
operations. But this should reduce the probability.

Longhorn #2278

Signed-off-by: Shuo Wu <shuo.wu@suse.com>