Add ensureMountPoint / cleanupMountPoint util functions
ensureMountPoint evaluates whether a path is a valid mountPoint when
detecting a corrupt mount point it will proceed with cleanup and return
an error. In the future we can look into successful remounts as an
alternative to returning an error here.

This would require NodeStageVolume then NodePublishVolume can just become
a bind mount and we layer it on top of the previous bind mount if it exists.

Longhorn #2629

Signed-off-by: Joshua Moody <joshua.moody@suse.com>