Delete CronJob instead of suspending
When volume is not attached and the setting `Allow Recurring Job While
Volume Is Detached` is disable, we delete the CronJob instead of suspending
it. When volume is attached, we recreate the CronJob. This is to prevent
Kubernetes from permanantly stopping the CronJob if the volume is detached
for a long time

Longhorn #2513

Signed-off-by: Phan Le <phan.le@rancher.com>