Add data source fields in BackingImageSpec and move the data source CR creation to backing-image-controller
If the owner of BackingImageDataSource is the user(HTTP API), it's
better not to update the CRs in backing-image-controller. Owner here
means the creator of a resource.

By moving creation of BackingImageDataSource CRs to the backing-
image-controller, the owners changes. Then it's fine for
the controller to update the CR specs. Besides, this simplify the
implementation of the backing image HTTP APIs, as well as guarantee
the existence of the BackingImageDataSource CR for each BackingImage.

Longhorn #2530

Signed-off-by: Shuo Wu <shuo.wu@suse.com>