Fix the rebuilding race condition
we use engine.Status.CurrentReplicaAddressMap to know which replica
needed to rebuild. Need to persist the engine.Status.CurrentReplicaAddressMap
to etcd before start rebuilding. Otherwise we might start rebuild and
later on engine.Status.CurrentReplicaAddressMap might be failed to persist
in etcd. In this case, we make the rebuilding replica as UNKNOWN and remove it
while it is in rebuilding

Longhorn #2849

Signed-off-by: Phan Le <phan.le@rancher.com>