Add support for encrypted RWX volumes
We retrieve the secret as part of the share-manager controller and
provide the passphrase as an environment variable to the new pod.

From a security perspective this isn't great, but for now since we don't
have a way to communicate with the share-manager pod this will have to do.

The alternative is to mount the secret via a kubernetes mechanism, but
this would only allow for using secrets in the longhorn-system namespace.
Further this would also mean that the share-manager requires knowledge
of our secret format/types for different keys.

Longhorn #2867

Signed-off-by: Joshua Moody <joshua.moody@suse.com>